+++ dca/tl_news.php
@@ -46,14 +46,30 @@ class tl_news_categories extends Backend
      */
     public function updateCategories(DataContainer $dc)
     {
-        $this->deleteCategories($dc);
+        $this->import('BackendUser', 'User');
         $arrCategories = deserialize($dc->activeRecord->categories);
+
+        // Use the default categories if the user is not allowed to edit the field directly
+        if (!$this->User->isAdmin && !in_array('tl_news::categories', $this->User->alexf)) {
+
+            // Return if the record is not new
+            if ($dc->activeRecord->tstamp) {
+                return;
+            }
+
+            $arrCategories = $this->User->newscategories_default;
+        }
+
+        $this->deleteCategories($dc);
 
         if (is_array($arrCategories) && !empty($arrCategories)) {
             foreach ($arrCategories as $intCategory) {
                 $this->Database->prepare("INSERT INTO tl_news_categories (category_id, news_id) VALUES (?, ?)")
                                ->execute($intCategory, $dc->id);
             }
+
+            $this->Database->prepare("UPDATE tl_news SET categories=? WHERE id=?")
+                           ->execute(serialize($arrCategories), $dc->id);
         }
     }