+++ dca/tl_user.php
@@ -15,8 +15,8 @@
 /**
  * Add a palette to tl_user
  */
-$GLOBALS['TL_DCA']['tl_user']['palettes']['extend'] = str_replace('newsfeedp;', 'newsfeedp,newscategories;', $GLOBALS['TL_DCA']['tl_user']['palettes']['extend']);
-$GLOBALS['TL_DCA']['tl_user']['palettes']['custom'] = str_replace('newsfeedp;', 'newsfeedp,newscategories;', $GLOBALS['TL_DCA']['tl_user']['palettes']['custom']);
+$GLOBALS['TL_DCA']['tl_user']['palettes']['extend'] = str_replace('newsfeedp;', 'newsfeedp;{newscategories_legend},newscategories,newscategories_default;', $GLOBALS['TL_DCA']['tl_user']['palettes']['extend']);
+$GLOBALS['TL_DCA']['tl_user']['palettes']['custom'] = str_replace('newsfeedp;', 'newsfeedp;{newscategories_legend},newscategories,newscategories_default;', $GLOBALS['TL_DCA']['tl_user']['palettes']['custom']);
 
 /**
  * Add a new field to tl_user
@@ -27,4 +27,14 @@ $GLOBALS['TL_DCA']['tl_user']['fields']['newscategories'] = array
     'exclude'                 => true,
     'inputType'               => 'checkbox',
     'sql'                     => "char(1) NOT NULL default ''"
+);
+
+$GLOBALS['TL_DCA']['tl_user']['fields']['newscategories_default'] = array
+(
+    'label'                   => &$GLOBALS['TL_LANG']['tl_user']['newscategories_default'],
+    'exclude'                 => true,
+    'inputType'               => 'treePicker',
+    'foreignKey'              => 'tl_news_category.title',
+    'eval'                    => array('multiple'=>true, 'fieldType'=>'checkbox', 'foreignTable'=>'tl_news_category', 'titleField'=>'title', 'searchField'=>'title', 'managerHref'=>'do=news&table=tl_news_category'),
+    'sql'                     => "blob NULL"
 );