+++ languages/en/tl_user.php
@@ -15,4 +15,10 @@
 /**
  * Fields
  */
-$GLOBALS['TL_LANG']['tl_user']['newscategories'] = array('Manage news categories', 'Allow the user to manage categories.');
+$GLOBALS['TL_LANG']['tl_user']['newscategories']         = array('Manage news categories', 'Allow the user to manage categories.');
+$GLOBALS['TL_LANG']['tl_user']['newscategories_default'] = array('Default news categories', 'Here you can set the default news categories that will be used if the user has no permissions to edit the categories field.');
+
+/**
+ * Legends
+ */
+$GLOBALS['TL_LANG']['tl_user']['newscategories_legend'] = 'News categories permissions';